# ClientApp/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Proiect_disertatie",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:5001"
    }
  ],
  "paths": {
...
    "/api/Movies/GetMoviesFor/{username}": {
      "get": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Movies/GetAllMovies": {
      "get": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "name": "SearchedMovie",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Movies/{title}": {
      "get": {
        "tags": [
          "Movies"
        ],
        "operationId": "GetMovie",
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Movies/GetReviewsFor/{movieId}": {
      "get": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Movies/AddMovieToUser/{movieId}": {
      "post": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Movies/AddReviewFor/{movieId}": {
      "post": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Movies/DeleteMovieFromUser/{movieId}": {
      "delete": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Movies/DeleteReviewFor/{movieId}": {
      "delete": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          {
            "name": "reviewId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TvShows/GetTvShowsFor/{username}": {
      "get": {
        "tags": [
          "TvShows"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TvShows/GetAllTvShows": {
      "get": {
        "tags": [
          "TvShows"
        ],
        "parameters": [
          {
            "name": "SearchedTvShow",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TvShows/{title}": {
      "get": {
        "tags": [
          "TvShows"
        ],
        "operationId": "GetTvShow",
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TvShow"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TvShow"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TvShow"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/TvShows/GetReviewsFor/{tvShowId}": {
      "get": {
        "tags": [
          "TvShows"
        ],
        "parameters": [
          {
            "name": "tvShowId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/TvShows/AddTvShowToUser/{tvShowId}": {
      "post": {
        "tags": [
          "TvShows"
        ],
        "parameters": [
          {
            "name": "tvShowId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TvShows/AddReviewFor/{tvShowId}": {
      "post": {
        "tags": [
          "TvShows"
        ],
        "parameters": [
          {
            "name": "tvShowId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TvShows/DeleteTvShowFromUser/{tvShowId}": {
      "delete": {
        "tags": [
          "TvShows"
        ],
        "parameters": [
          {
            "name": "tvShowId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TvShows/DeleteReviewFor/{tvShowId}": {
      "delete": {
        "tags": [
          "TvShows"
        ],
        "parameters": [
          {
            "name": "tvShowId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          {
            "name": "reviewId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActorList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "asCharacter": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CompanyList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CountryList": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DirectorList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GenreList": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LanguageList": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "movieId": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "originalTitle": {
            "type": "string",
            "nullable": true
          },
          "fullTitle": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "nullable": true
          },
          "runtimeMins": {
            "type": "string",
            "nullable": true
          },
          "runtimeStr": {
            "type": "string",
            "nullable": true
          },
          "plot": {
            "type": "string",
            "nullable": true
          },
          "plotLocal": {
            "type": "string",
            "nullable": true
          },
          "plotLocalIsRtl": {
            "type": "boolean",
            "nullable": true
          },
          "awards": {
            "type": "string",
            "nullable": true
          },
          "directors": {
            "type": "string",
            "nullable": true
          },
          "directorList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DirectorList"
            },
            "nullable": true
          },
          "writers": {
            "type": "string",
            "nullable": true
          },
          "writerList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WriterList"
            },
            "nullable": true
          },
          "stars": {
            "type": "string",
            "nullable": true
          },
          "starList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StarList"
            },
            "nullable": true
          },
          "actorList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActorList"
            },
            "nullable": true
          },
          "genres": {
            "type": "string",
            "nullable": true
          },
          "genreList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenreList"
            },
            "nullable": true
          },
          "companies": {
            "type": "string",
            "nullable": true
          },
          "companyList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyList"
            },
            "nullable": true
          },
          "countries": {
            "type": "string",
            "nullable": true
          },
          "countryList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CountryList"
            },
            "nullable": true
          },
          "languages": {
            "type": "string",
            "nullable": true
          },
          "languageList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LanguageList"
            },
            "nullable": true
          },
          "contentRating": {
            "type": "string",
            "nullable": true
          },
          "imDbRating": {
            "type": "string",
            "nullable": true
          },
          "imDbRatingVotes": {
            "type": "string",
            "nullable": true
          },
          "metacriticRating": {
            "type": "string",
            "nullable": true
          },
          "ratings": {
            "$ref": "#/components/schemas/Ratings"
          },
          "wikipedia": {
            "$ref": "#/components/schemas/Wikipedia"
          },
          "images": {
            "$ref": "#/components/schemas/MoviesImages"
          },
          "trailer": {
            "$ref": "#/components/schemas/MoviesTrailer"
          },
          "boxOffice": {
            "$ref": "#/components/schemas/BoxOffice"
          },
          "tagline": {
            "type": "string",
            "nullable": true
          },
          "keywords": {
            "type": "string",
            "nullable": true
          },
          "similars": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Similar"
            },
            "nullable": true
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ObjectId": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "machine": {
            "type": "integer",
            "format": "int32",
            "readOnly": true,
            "deprecated": true
          },
          "pid": {
            "type": "integer",
            "format": "int32",
            "readOnly": true,
            "deprecated": true
          },
          "increment": {
            "type": "integer",
            "format": "int32",
            "readOnly": true,
            "deprecated": true
          },
          "creationTime": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ReviewDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "short_description": {
            "type": "string",
            "nullable": true
          },
          "main_description": {
            "type": "string",
            "nullable": true
          },
          "stars": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "StarList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TvShow": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "tvShow_Id": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "originalTitle": {
            "type": "string",
            "nullable": true
          },
          "fullTitle": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "nullable": true
          },
          "runtimeMins": {
            "type": "string",
            "nullable": true
          },
          "runtimeStr": {
            "type": "string",
            "nullable": true
          },
          "plot": {
            "type": "string",
            "nullable": true
          },
          "plotLocal": {
            "type": "string",
            "nullable": true
          },
          "plotLocalIsRtl": {
            "type": "boolean",
            "nullable": true
          },
          "awards": {
            "type": "string",
            "nullable": true
          },
          "directors": {
            "type": "string",
            "nullable": true
          },
          "writers": {
            "type": "string",
            "nullable": true
          },
          "stars": {
            "type": "string",
            "nullable": true
          },
          "starList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TvShowStarList"
            },
            "nullable": true
          },
          "actorList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TvShowActorList"
            },
            "nullable": true
          },
          "genres": {
            "type": "string",
            "nullable": true
          },
          "genreList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TvShowGenreList"
            },
            "nullable": true
          },
          "companies": {
            "type": "string",
            "nullable": true
          },
          "companyList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TvShowCompanyList"
            },
            "nullable": true
          },
          "countries": {
            "type": "string",
            "nullable": true
          },
          "countryList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TvShowCountryList"
            },
            "nullable": true
          },
          "languages": {
            "type": "string",
            "nullable": true
          },
          "languageList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TvShowLanguageList"
            },
            "nullable": true
          },
          "contentRating": {
            "type": "string",
            "nullable": true
          },
          "imDbRating": {
            "type": "string",
            "nullable": true
          },
          "imDbRatingVotes": {
            "type": "string",
            "nullable": true
          },
          "metacriticRating": {
            "type": "string",
            "nullable": true
          },
          "ratings": {
            "$ref": "#/components/schemas/TvShowRatings"
          },
          "wikipedia": {
            "$ref": "#/components/schemas/TvShowWikipedia"
          },
          "images": {
            "$ref": "#/components/schemas/TvShowImages"
          },
          "trailer": {
            "$ref": "#/components/schemas/TvShowTrailer"
          },
          "boxOffice": {
            "$ref": "#/components/schemas/TvShowBoxOffice"
          },
          "keywords": {
            "type": "string",
            "nullable": true
          },
          "similars": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TvShowSimilar"
            },
            "nullable": true
          },
          "tvSeriesInfo": {
            "$ref": "#/components/schemas/TvSeriesInfo"
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WriterList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}